Cleaning up scripts for testing on Warwick systems with updated Intel and PGI compilers

+++ apps/c/mb_shsgc/Max_datatransfer/CUDA/gridgen_kernel_cuda_kernel.cu
@@ -104,15 +104,14 @@ void ops_par_loop_gridgen_kernel_execute(ops_kernel_descriptor *desc) {
   dim3 grid( (x_size-1)/block->instance->OPS_block_size_x+ 1, 1, 1);
   dim3 tblock(block->instance->OPS_block_size_x,1,1);
 
-
-
-  int dat0 = (block->instance->OPS_soa ? args[0].dat->type_size : args[0].dat->elem_size);
+  long long int dat0 = (block->instance->OPS_soa ? args[0].dat->type_size
+                                                 : args[0].dat->elem_size);
 
   char *p_a[2];
 
   //set up initial pointers
-  int base0 = args[0].dat->base_offset + 
-           dat0 * 1 * (start[0] * args[0].stencil->stride[0]);
+  long long int base0 = args[0].dat->base_offset +
+                        dat0 * 1 * (start[0] * args[0].stencil->stride[0]);
   p_a[0] = (char *)args[0].data_d + base0;
