+++ apps/c/shsgc/CUDA/Riemann_kernel_cuda_kernel.cu
@@ -178,40 +178,44 @@ void ops_par_loop_Riemann_kernel_execute(ops_kernel_descriptor *desc) {
   dim3 grid( (x_size-1)/block->instance->OPS_block_size_x+ 1, 1, 1);
   dim3 tblock(block->instance->OPS_block_size_x,1,1);
 
-
-
-  int dat0 = (block->instance->OPS_soa ? args[0].dat->type_size : args[0].dat->elem_size);
-  int dat1 = (block->instance->OPS_soa ? args[1].dat->type_size : args[1].dat->elem_size);
-  int dat2 = (block->instance->OPS_soa ? args[2].dat->type_size : args[2].dat->elem_size);
-  int dat3 = (block->instance->OPS_soa ? args[3].dat->type_size : args[3].dat->elem_size);
-  int dat4 = (block->instance->OPS_soa ? args[4].dat->type_size : args[4].dat->elem_size);
-  int dat5 = (block->instance->OPS_soa ? args[5].dat->type_size : args[5].dat->elem_size);
+  long long int dat0 = (block->instance->OPS_soa ? args[0].dat->type_size
+                                                 : args[0].dat->elem_size);
+  long long int dat1 = (block->instance->OPS_soa ? args[1].dat->type_size
+                                                 : args[1].dat->elem_size);
+  long long int dat2 = (block->instance->OPS_soa ? args[2].dat->type_size
+                                                 : args[2].dat->elem_size);
+  long long int dat3 = (block->instance->OPS_soa ? args[3].dat->type_size
+                                                 : args[3].dat->elem_size);
+  long long int dat4 = (block->instance->OPS_soa ? args[4].dat->type_size
+                                                 : args[4].dat->elem_size);
+  long long int dat5 = (block->instance->OPS_soa ? args[5].dat->type_size
+                                                 : args[5].dat->elem_size);
 
   char *p_a[6];
 
   //set up initial pointers
-  int base0 = args[0].dat->base_offset + 
-           dat0 * 1 * (start[0] * args[0].stencil->stride[0]);
+  long long int base0 = args[0].dat->base_offset +
+                        dat0 * 1 * (start[0] * args[0].stencil->stride[0]);
   p_a[0] = (char *)args[0].data_d + base0;
 
-  int base1 = args[1].dat->base_offset + 
-           dat1 * 1 * (start[0] * args[1].stencil->stride[0]);
+  long long int base1 = args[1].dat->base_offset +
+                        dat1 * 1 * (start[0] * args[1].stencil->stride[0]);
   p_a[1] = (char *)args[1].data_d + base1;
 
-  int base2 = args[2].dat->base_offset + 
-           dat2 * 1 * (start[0] * args[2].stencil->stride[0]);
+  long long int base2 = args[2].dat->base_offset +
+                        dat2 * 1 * (start[0] * args[2].stencil->stride[0]);
   p_a[2] = (char *)args[2].data_d + base2;
 
-  int base3 = args[3].dat->base_offset + 
-           dat3 * 1 * (start[0] * args[3].stencil->stride[0]);
+  long long int base3 = args[3].dat->base_offset +
+                        dat3 * 1 * (start[0] * args[3].stencil->stride[0]);
   p_a[3] = (char *)args[3].data_d + base3;
 
-  int base4 = args[4].dat->base_offset + 
-           dat4 * 1 * (start[0] * args[4].stencil->stride[0]);
+  long long int base4 = args[4].dat->base_offset +
+                        dat4 * 1 * (start[0] * args[4].stencil->stride[0]);
   p_a[4] = (char *)args[4].data_d + base4;
 
-  int base5 = args[5].dat->base_offset + 
-           dat5 * 1 * (start[0] * args[5].stencil->stride[0]);
+  long long int base5 = args[5].dat->base_offset +
+                        dat5 * 1 * (start[0] * args[5].stencil->stride[0]);
   p_a[5] = (char *)args[5].data_d + base5;
