+++ apps/c/shsgc/CUDA/test_kernel_cuda_kernel.cu
@@ -126,13 +126,14 @@ void ops_par_loop_test_kernel_execute(ops_kernel_descriptor *desc) {
 
 
   mvReductArraysToDevice(block->instance,reduct_bytes);
-  int dat0 = (block->instance->OPS_soa ? args[0].dat->type_size : args[0].dat->elem_size);
+  long long int dat0 = (block->instance->OPS_soa ? args[0].dat->type_size
+                                                 : args[0].dat->elem_size);
 
   char *p_a[2];
 
   //set up initial pointers
-  int base0 = args[0].dat->base_offset + 
-           dat0 * 1 * (start[0] * args[0].stencil->stride[0]);
+  long long int base0 = args[0].dat->base_offset +
+                        dat0 * 1 * (start[0] * args[0].stencil->stride[0]);
   p_a[0] = (char *)args[0].data_d + base0;
